+++ CHANGELOG.md
@@ -804,3 +804,4 @@
 - [2026-05-02 20:56] perf(theme): optimize memory consumption in theme list viewport
 - [2026-05-02 21:56] style(network): harmonize light and dark theme palette for network
 - [2026-05-02 22:13] test(forms): add unit tests for forms business logic
+- [2026-05-02 22:23] test(network): add unit tests for network business logic